Fix arm build.

+++ git.h
@@ -110,7 +110,7 @@ struct Object {
 		(b)[7] = (n) >> 0; \
 	} while(0)
 
-#define QDIR(qid)	((qid)->path & (0xff))
+#define QDIR(qid)	((int)(qid)->path & (0xff))
 #define QPATH(id, dt)	(((id) << 8) | ((dt) & 0xff))
 #define isblank(c) \
 	(((c) != '\n') && isspace(c))
